backend/apk2img_tool/requirement.txt
pandas
numpy
scikit-learn
matplotlib
androguard
Pillow